requirements.txt
streamlit
numpy
pandas
opencv-python
tensorflow
scikit-learn
Pillow